+++ RELEASE-NOTES.md
@@ -7,7 +7,7 @@
 * test: 测试 case 优化。
 * refactor: token失效或者没有token的情况下，调接口应该emit error。
 * refactor: importUser, generalSearchUsers,createDrive 3个接口新增一些可选参数。
-
+* fix: 修复httpClient，以支持分享页面同时使用 x-share-token和access token的场景。
 
 ## 1.1.0